requirements.txt
gunicorn==20.0.4
whitenoise==5.3.0
Django==3.1.4
django-heroku==0.3.1
django-allauth==0.44.0
django-filter==21.1
djangorestframework==3.12.4
django-bootstrap4==2.3.1
dj-database-url==0.5.0
django-import-export==2.7.0
pandas==1.1.5
PyJWT==2.0.0
